predict4java azimuths: widen Int -> Double at the bridge
getAosAzimuth() / getLosAzimuth() return int (integer degrees), not double
like getMaxEl(). The SatellitePass domain type stores Doubles so other
satellite sources we might wire in later can carry sub-degree precision.
Add the .toDouble() cast at the predict4java seam.

+++ app/src/main/java/de/astronarren/allsky/data/astro/SatelliteRepository.kt
@@ -120,8 +120,10 @@ class SatelliteRepository(
             maxElevationDeg = getMaxEl(),
             // predict4java names rise/set azimuths after the amateur-radio
             // AOS (Acquisition of Signal) / LOS (Loss of Signal) convention.
-            startAzimuthDeg = getAosAzimuth(),
-            endAzimuthDeg = getLosAzimuth(),
+            // Returned as integer degrees — widen to Double for our domain
+            // type, which keeps the door open for finer-grained sources later.
+            startAzimuthDeg = getAosAzimuth().toDouble(),
+            endAzimuthDeg = getLosAzimuth().toDouble(),
         )
     }
